ORM relation and bugfix

+++ database/migrations/2019_02_17_194044_create_products_table.php
@@ -14,14 +14,14 @@ class CreateProductsTable extends Migration
     public function up()
     {
         Schema::create('products', function (Blueprint $table) {
-            
+
             $table->increments('id');
 
             $table->string('name');
 
             $table->text('detail');
 
-            $table->double('price');
+            $table->integer('price');
 
             $table->integer('stock');
